Create packages duke.ui, duke.component, duke.excpetions, duke.command, duke.task, Move code into respective packages
