+++ src/main/java/duke/exception/InvalidInputException.java
@@ -1,0 +1,6 @@
+package duke.exception;
+
+import duke.ui.Luke;
+
+// exceptions
+public class InvalidInputException extends Exception {}
